requirements.txt
certifi==2021.10.8
numpy==1.21.4
Pillow==8.4.0
torch==1.10.0+cu102
torchaudio==0.10.0+cu102
torchvision==0.11.1+cu102
typing-extensions==3.10.0.2
wincertstore==0.2
PyYAML=6.0
pandas=1.3.4
opencv-contrib-python=4.5.4.60
opencv-python=4.5.4.60